separate lambda functions in handlers, removed un-used/extra code
