--- src/lambdas/types.ts
@@ -1,5 +1,0 @@
-// types specific to function handlers
-
-export interface Response {
-  statusCode: number
-}
